# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LoyalBox API",
    "version": "1.0.0",
    "description": "API for LoyalBox loyalty program, allowing users to register, log in, earn/redeem points, and admins to manage rewards and users."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "points": { "type": "number" },
          "isAdmin": { "type": "boolean" }
        }
      },
      "Reward": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "pointsRequired": { "type": "number" },
          "description": { "type": "string" }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "type": { "type": "string", "enum": ["earn", "redeem"] },
          "points": { "type": "number" },
          "description": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "testuser" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request (e.g., missing fields or username exists)" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in a user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "testuser" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid credentials" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/auth/promote/{id}": {
      "put": {
        "summary": "Promote a user to admin (admin only)",
        "tags": ["Auth"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "User ID to promote"
          }
        ],
        "responses": {
          "200": {
            "description": "User promoted to admin",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "User promoted to admin" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized (invalid or missing token)" },
          "403": { "description": "Forbidden (requires admin access)" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/rewards": {
      "get": {
        "summary": "Get all rewards",
        "tags": ["Rewards"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of rewards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Reward" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "summary": "Create a new reward (admin only)",
        "tags": ["Rewards"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Gift Card" },
                  "pointsRequired": { "type": "number", "example": 100 },
                  "description": { "type": "string", "example": "$10 Gift Card" }
                },
                "required": ["name", "pointsRequired"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reward created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Reward" }
              }
            }
          },
          "400": { "description": "Bad request (missing fields)" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden (requires admin access)" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/rewards/{id}": {
      "put": {
        "summary": "Update a reward (admin only)",
        "tags": ["Rewards"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Reward ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Gift Card" },
                  "pointsRequired": { "type": "number", "example": 100 },
                  "description": { "type": "string", "example": "$10 Gift Card" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reward updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Reward" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden (requires admin access)" },
          "404": { "description": "Reward not found" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "summary": "Delete a reward (admin only)",
        "tags": ["Rewards"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Reward ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Reward deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Reward deleted" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden (requires admin access)" },
          "404": { "description": "Reward not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/transactions/buy": {
      "post": {
        "summary": "Buy an item and earn points",
        "tags": ["Transactions"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "itemName": { "type": "string", "example": "Shirt" },
                  "price": { "type": "number", "example": 50 }
                },
                "required": ["itemName", "price"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Points earned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Points earned" },
                    "points": { "type": "number" }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request (missing fields)" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/transactions/redeem/{rewardId}": {
      "post": {
        "summary": "Redeem points for a reward",
        "tags": ["Transactions"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "rewardId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Reward ID to redeem"
          }
        ],
        "responses": {
          "200": {
            "description": "Reward redeemed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Reward redeemed" },
                    "points": { "type": "number" }
                  }
                }
              }
            }
          },
          "400": { "description": "Insufficient points" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Reward not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/transactions/history": {
      "get": {
        "summary": "View personal transaction history",
        "tags": ["Transactions"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Transaction" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/transactions/users": {
      "get": {
        "summary": "View all users and their points (admin only)",
        "tags": ["Transactions"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden (requires admin access)" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/transactions/all-transactions": {
      "get": {
        "summary": "View all transactions (admin only)",
        "tags": ["Transactions"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of all transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Transaction" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden (requires admin access)" },
          "500": { "description": "Server error" }
        }
      }
    }
  }
}